Fix some conformance-based compiler warnings

+++ MOULOpenSourceClientPlugin/Plasma20/Sources/Plasma/FeatureLib/pfGameMgr/BlueSpiral/pfGmBlueSpiral.cpp
@@ -253,7 +253,7 @@ void pfGmBlueSpiral::HitCloth (int clothNum) {
 	msg.messageBytes	= sizeof(msg);
 	msg.recvGameId		= GetGameId();	// send to GameSrv on server
 	msg.transId			= 0;
-	msg.clothNum		= clothNum;
+	msg.clothNum		= (byte)clothNum;
 
 	GameMgrSend(&msg);
 }
